+++ Tests/Subscribe.CleanArchDapr.TestApplication/Subscribe.CleanArchDapr.TestApplication.Api/Controllers/DaprEventHandlerController.cs
@@ -1,10 +1,13 @@
+using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Dapr;
 using Intent.RoslynWeaver.Attributes;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.DependencyInjection;
 using Publish.CleanArchDapr.TestApplication.Eventing.Messages;
+using Subscribe.CleanArchDapr.TestApplication.Application.Common.Eventing;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: DefaultIntentManaged(Mode.Fully, Targets = Targets.Usings)]
@@ -17,17 +20,20 @@ namespace Subscribe.CleanArchDapr.TestApplication.Api.Controllers
     public class DaprEventHandlerController : ControllerBase
     {
         private readonly ISender _mediatr;
+        private readonly IServiceProvider _serviceProvider;
 
-        public DaprEventHandlerController(ISender mediatr)
+        public DaprEventHandlerController(ISender mediatr, IServiceProvider serviceProvider)
         {
             _mediatr = mediatr;
+            _serviceProvider = serviceProvider;
         }
 
         [HttpPost]
         [Topic(CustomerCreatedEvent.PubsubName, CustomerCreatedEvent.TopicName)]
         public async Task HandleCustomerCreatedEvent(CustomerCreatedEvent @event, CancellationToken cancellationToken)
         {
-            await _mediatr.Send(@event, cancellationToken);
+            var handler = _serviceProvider.GetRequiredService<IIntegrationEventHandler<CustomerCreatedEvent>>();
+            await handler.HandleAsync(@event, cancellationToken);
         }
 
         [HttpPost]